refactor(#3274): extract object-runtime descriptor/integrity helpers (WAVE-B slice 1)
Decompose the ~7,378-LOC `ensureObjectRuntime` in object-runtime.ts by lifting
its property-descriptor + object-integrity helper-build block (~2,464 LOC)
VERBATIM into a new sibling module `object-runtime-descriptors.ts`
(`buildObjectDescriptorHelpers`), replacing the inline block with one call that
threads the captured scope (the `registerNative` minter, object-runtime type
indices / ValType aliases, dependency func indices, and $PropEntry.$flags /
$Object.flags bit constants) through a typed `ObjectDescriptorHelperState`
bundle. Because the moved code is character-for-character identical and the
same `registerNative` closure runs the same blocks in the same order, the
minted func-index sequence — and the emitted Wasm — is preserved exactly.

Helpers relocated: __defineProperty_value/_accessor, __defineProperties,
__obj_define_from_desc, __getOwnPropertyDescriptor, __create_descriptor,
__create_accessor_descriptor, __getOwnPropertyNames/__getOwnPropertySymbols,
__object_getOwnPropertyDescriptors/__object_fromEntries, and the integrity
predicate + set-path helpers (isFrozen/isSealed/isExtensible,
preventExtensions/seal/freeze).

object-runtime.ts: 10,149 -> 7,721 LOC; ensureObjectRuntime: ~7,378 -> ~4,950.

Gates: `prove-emit-identity check` -> IDENTICAL (39/39 gc/standalone/wasi);
`tsc --noEmit` -> 0; `tests/issue-3274.test.ts` (4) green — descriptor/accessor/
gopd helpers reachable + correct under --target standalone, zero host imports.

Subtask of #3182 (code-bloat elimination epic).

✓
